Invert tests wrt file/inline config
Previously the initial setup made a file that housed the config; this
requires a lot more ceremony in each test and distances task config from
pipeline invocation of the task too much.

Instead, use inline configs to make it very clear what's going on. This
also means we no longer need to specify params that the file-based task
config required in other tests.